requirements.txt
absl-py==0.5.0
antspy==0.1.4
apipkg==1.4
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.7.1
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
bleach==1.5.0
boto==2.48.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
citeproc-py==0.4.0
click==6.7
cloudpickle==0.7.0
conda==4.4.11
constantly==15.1.0
coverage==3.7.1
cryptography==2.1.4
cssselect==1.0.3
cycler==0.10.0
dask==1.1.1
datalad==0.9.2
decorator==4.2.1
duecredit==0.6.3
entrypoints==0.2.3
et-xmlfile==1.0.1
execnet==1.5.0
funcsigs==1.0.2
future==0.16.0
gast==0.2.0
gitdb2==2.0.3
GitPython==2.1.8
grabbit==0.1.1
graphviz==0.10.1
grpcio==1.15.0
h5py==2.7.1
html5lib==0.9999999
httpretty==0.8.14
humanize==0.5.1
hyperlink==18.0.0
hyperopt==0.1.1
idna==2.6
incremental==17.5.0
ipdb==0.11
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
iso8601==0.1.12
isodate==0.6.0
jdcal==1.4
jedi==0.11.1
Jinja2==2.10
joblib==0.13.0
jsmin==2.2.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.10
jupyterlab-launcher==0.10.5
Keras==2.1.4
keyring==11.0.0
keyrings.alt==3.0
lxml==4.1.1
Markdown==3.0
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
mock==2.0.0
morphsnakes==2.0.1
mpmath==1.0.0
msgpack==0.5.6
multidict==4.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nibabel==2.2.1
nilearn==0.4.0
nipy==0.4.2
nipype==1.0.2.dev0
nose==1.3.7
notebook==5.4.0
numpy==1.14.5
olefile==0.45.1
openpyxl==2.5.11
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
parsel==1.4.0
parso==0.1.1
patool==1.12
patsy==0.5.0
pbr==3.1.1
pep8==1.7.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
pluggy==0.6.0
prompt-toolkit==1.0.15
protobuf==3.6.1
prov==1.5.0
ptyprocess==0.5.2
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybids==0.4.2
pycosat==0.6.3
pycparser==2.18
PyDispatcher==2.0.5
pydot==1.2.4
pydot-ng==1.0.0
pydotplus==2.0.2
PyGithub==1.37
Pygments==2.2.0
PyJWT==1.6.0
PyLD==0.8.2
pymongo==3.7.2
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.4.2
pytest-cache==1.0
pytest-cov==2.5.1
pytest-forked==0.2
pytest-pep8==1.0.6
pytest-xdist==1.22.2
python-coveralls==2.9.1
python-dateutil==2.6.1
pytz==2018.3
PyWavelets==1.0.1
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
queuelib==1.4.2
rdflib==4.2.2
requests==2.18.4
requests-ftp==0.3.1
ruamel-yaml==0.15.35
scikit-image==0.14.2
scikit-learn==0.19.1
scipy==1.0.0
Scrapy==1.5.0
seaborn==0.8.1
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
SimpleITK==1.1.0
simplejson==3.13.2
six==1.11.0
smmap2==2.0.3
statsmodels==0.8.0
sympy==1.1.1
tensorboard==1.10.0
tensorflow-gpu==1.10.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
torchsummary==1.5.1
tornado==4.5.3
tqdm==4.19.6
traitlets==4.3.2
traits==4.6.0
Twisted==17.9.0
urllib3==1.22
vcrpy==1.11.1
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4
wrapt==1.10.11
yarl==1.1.1
zope.interface==4.4.3
